requirements.txt
numpy~=2.0.2
tensorflow~=2.18.0
flask~=3.1.0
pillow~=11.0.0